Move vch appliance detection to general.component
