+++ h5c/vic/src/vic-webapp/src/app/create-vch-wizard/summary/summary.html
@@ -1,12 +1,3 @@
-<clr-alert *ngIf="isApplianceUnavailable"
-           [clrAlertType]="'alert-danger'"
-           [clrAlertClosable]="false">
-  <div class="alert-item">
-    <div class="alert-text">
-      VIC appliance VM was not found or unreachable. Please make sure you have deployed it correctly.
-    </div>
-  </div>
-</clr-alert>
 <form [formGroup]="form" novalidate>
   <section class="form-block">
     <div class="form-group row">
